requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
djangorestframework==3.15.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
oauthlib==3.2.2
propcache==0.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
twilio==9.4.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
